bo7: wire psp boss floor snap as C in mixed TU (parametrized arena floor)
